[kernel/refactor] Move threading stuff to own directory
