+++ src/emulator/host/src/host.cpp
@@ -25,7 +25,7 @@
 #include <audio/functions.h>
 #include <io/functions.h>
 #include <kernel/functions.h>
-#include <kernel/thread_state.h>
+#include <kernel/thread/thread_state.h>
 #include <nids/functions.h>
 #include <util/lock_and_find.h>
 #include <util/log.h>
