+++ src/emulator/kernel/include/kernel/thread/thread_functions.h
