+++ src/emulator/kernel/src/thread/thread.cpp
