+++ src/emulator/modules/SceGxm/SceGxm.cpp
@@ -25,7 +25,7 @@
 
 #include <cpu/functions.h>
 #include <host/functions.h>
-#include <kernel/thread_functions.h>
+#include <kernel/thread/thread_functions.h>
 #include <psp2/kernel/error.h>
 
 #include <algorithm>
