+++ src/emulator/modules/SceLibKernel/SceLibKernel.cpp
@@ -21,7 +21,7 @@
 #include <host/functions.h>
 #include <io/functions.h>
 #include <kernel/functions.h>
-#include <kernel/thread_functions.h>
+#include <kernel/thread/thread_functions.h>
 #include <rtc/rtc.h>
 #include <util/log.h>
